check update manually for portable mode
https://github.com/Wox-launcher/Wox/issues/2933
