+++ Wox.Core/Updater.cs
@@ -74,18 +74,7 @@ namespace Wox.Core
 
                     await updateManager.ApplyReleases(newUpdateInfo);
 
-                    if (DataLocation.PortableDataLocationInUse())
-                    {
-                        var targetDestination = updateManager.RootAppDirectory + $"\\app-{newReleaseVersion.ToString()}\\{DataLocation.PortableFolderName}";
-                        FilesFolders.Copy(DataLocation.PortableDataPath, targetDestination);
-                        if (!FilesFolders.VerifyBothFolderFilesEqual(DataLocation.PortableDataPath, targetDestination))
-                            MessageBox.Show(string.Format("Wox was not able to move your user profile data to the new update version. Please manually" +
-                                "move your profile data folder from {0} to {1}", DataLocation.PortableDataPath, targetDestination));
-                    }
-                    else
-                    {
-                        await updateManager.CreateUninstallerRegistryEntry();
-                    }
+                    await updateManager.CreateUninstallerRegistryEntry();
 
                     var newVersionTips = NewVersinoTips(newReleaseVersion.ToString());
 
@@ -124,11 +113,12 @@ namespace Wox.Core
             var json = await Http.Get(api);
 
             var releases = JsonConvert.DeserializeObject<List<GithubRelease>>(json).AsEnumerable();
-            if (!updateToPrereleases) {
+            if (!updateToPrereleases)
+            {
                 releases = releases.Where(r => !r.Prerelease);
             }
             var latest = releases.OrderByDescending(r => r.PublishedAt).First();
-            
+
             var latestUrl = latest.HtmlUrl.Replace("/tag/", "/download/");
 
             var client = new WebClient { Proxy = Http.WebProxy() };
